+++ src/mmw/apps/modeling/urls.py
@@ -21,21 +21,9 @@ urlpatterns = patterns(
     url(r'projects/(?P<proj_id>[0-9]+)$', views.project, name='project'),
     url(r'scenarios/$', views.scenarios, name='scenarios'),
     url(r'scenarios/(?P<scen_id>[0-9]+)$', views.scenario, name='scenario'),
-    url(r'start/analyze/land/$', views.start_analyze_land,
-        name='start_analyze_land'),
-    url(r'start/analyze/soil/$', views.start_analyze_soil,
-        name='start_analyze_soil'),
-    url(r'start/analyze/animals/$', views.start_analyze_animals,
-        name='start_analyze_animals'),
-    url(r'start/analyze/pointsource/$', views.start_analyze_pointsource,
-        name='start_analyze_pointsource'),
-    url(r'start/analyze/catchment-water-quality/$',
-        views.start_analyze_catchment_water_quality,
-        name='start_analyze_catchment_water_quality'),
     url(r'start/mapshed/$', views.start_mapshed, name='start_mapshed'),
     url(r'jobs/' + uuid_regex, views.get_job, name='get_job'),
     url(r'start/tr55/$', views.start_tr55, name='start_tr55'),
-    url(r'start/rwd/$', views.start_rwd, name='start_rwd'),
     url(r'start/gwlfe/$', views.start_gwlfe, name='start_gwlfe'),
     url(r'boundary-layers/(?P<table_code>\w+)/(?P<obj_id>[0-9]+)/$',
         views.boundary_layer_detail, name='boundary_layer_detail'),
