+++ src/mmw/apps/modeling/views.py
@@ -18,7 +18,8 @@ from django.conf import settings
 from django.db import connection
 from django.contrib.gis.geos import WKBReader
 from django.http import (HttpResponse,
-                         Http404)
+                         Http404,
+                         )
 from django.core.servers.basehttp import FileWrapper
 
 from apps.core.models import Job
@@ -158,38 +159,6 @@ def scenario(request, scen_id):
 
 @decorators.api_view(['POST'])
 @decorators.permission_classes((AllowAny, ))
-def start_rwd(request, format=None):
-    """
-    Starts a job to run Rapid Watershed Delineation on a point-based location.
-    """
-    user = request.user if request.user.is_authenticated() else None
-    created = now()
-    location = request.POST['location']
-    data_source = request.POST.get('dataSource', 'drb')
-
-    # Parse out the JS style T/F to a boolean
-    snappingParam = request.POST['snappingOn']
-    snapping = True if snappingParam == 'true' else False
-
-    job = Job.objects.create(created_at=created, result='', error='',
-                             traceback='', user=user, status='started')
-
-    task_list = _initiate_rwd_job_chain(location, snapping, data_source,
-                                        job.id)
-
-    job.uuid = task_list.id
-    job.save()
-
-    return Response(
-        {
-            'job': task_list.id,
-            'status': 'started',
-        }
-    )
-
-
-@decorators.api_view(['POST'])
-@decorators.permission_classes((AllowAny, ))
 def start_gwlfe(request, format=None):
     """
     Starts a job to run GWLF-E.
@@ -279,109 +248,6 @@ def export_gms(request, format=None):
     response['Content-Disposition'] = 'attachment; '\
                                       'filename={}.gms'.format(filename)
     return response
-
-
-@decorators.api_view(['POST'])
-@decorators.permission_classes((AllowAny, ))
-def start_analyze_land(request, format=None):
-    user = request.user if request.user.is_authenticated() else None
-
-    area_of_interest, wkaoi = parse_input(request.POST['analyze_input'])
-
-    geop_input = {'polygon': [area_of_interest]}
-
-    return start_celery_job([
-        geoprocessing.run.s('nlcd', geop_input, wkaoi),
-        tasks.analyze_nlcd.s(area_of_interest)
-    ], area_of_interest, user)
-
-
-@decorators.api_view(['POST'])
-@decorators.permission_classes((AllowAny, ))
-def start_analyze_soil(request, format=None):
-    user = request.user if request.user.is_authenticated() else None
-
-    area_of_interest, wkaoi = parse_input(request.POST['analyze_input'])
-
-    geop_input = {'polygon': [area_of_interest]}
-
-    return start_celery_job([
-        geoprocessing.run.s('soil', geop_input, wkaoi),
-        tasks.analyze_soil.s(area_of_interest)
-    ], area_of_interest, user)
-
-
-@decorators.api_view(['POST'])
-@decorators.permission_classes((AllowAny, ))
-def start_analyze_animals(request, format=None):
-    user = request.user if request.user.is_authenticated() else None
-    area_of_interest, __ = parse_input(request.POST['analyze_input'])
-
-    return start_celery_job([
-        tasks.analyze_animals.s(area_of_interest)
-    ], area_of_interest, user)
-
-
-@decorators.api_view(['POST'])
-@decorators.permission_classes((AllowAny, ))
-def start_analyze_pointsource(request, format=None):
-    user = request.user if request.user.is_authenticated() else None
-    area_of_interest, __ = parse_input(request.POST['analyze_input'])
-
-    return start_celery_job([
-        tasks.analyze_pointsource.s(area_of_interest)
-    ], area_of_interest, user)
-
-
-@decorators.api_view(['POST'])
-@decorators.permission_classes((AllowAny, ))
-def start_analyze_catchment_water_quality(request, format=None):
-    user = request.user if request.user.is_authenticated() else None
-    area_of_interest, __ = parse_input(request.POST['analyze_input'])
-
-    return start_celery_job([
-        tasks.analyze_catchment_water_quality.s(area_of_interest)
-    ], area_of_interest, user)
-
-
-@decorators.api_view(['GET'])
-@decorators.permission_classes((AllowAny, ))
-def get_job(request, job_uuid, format=None):
-    # TODO consider if we should have some sort of session id check to ensure
-    # you can only view your own jobs.
-    try:
-        job = Job.objects.get(uuid=job_uuid)
-    except Job.DoesNotExist:
-        raise Http404("Not found.")
-
-    # Get the user so that logged in users can only see jobs that they started
-    # or anonymous ones
-    user = request.user if request.user.is_authenticated() else None
-
-    if job.user and job.user != user:
-        raise Http404("Not found.")
-
-    # TODO Should we return the error? Might leak info about the internal
-    # workings that we don't want exposed.
-    return Response(
-        {
-            'job_uuid': job.uuid,
-            'status': job.status,
-            'result': job.result,
-            'error': job.error,
-            'started': job.created_at,
-            'finished': job.delivered_at,
-        }
-    )
-
-
-def _initiate_rwd_job_chain(location, snapping, data_source,
-                            job_id, testing=False):
-    errback = save_job_error.s(job_id)
-
-    return chain(tasks.start_rwd_job.s(location, snapping, data_source),
-                 save_job_result.s(job_id, location)) \
-        .apply_async(link_error=errback)
 
 
 @decorators.api_view(['POST'])
@@ -612,35 +478,33 @@ def drb_point_sources(request):
                     headers={'Cache-Control': 'max-age: 604800'})
 
 
-def start_celery_job(task_list, job_input, user=None):
-    """
-    Given a list of Celery tasks and it's input, starts a Celery async job with
-    those tasks, adds save_job_result and save_job_error handlers, and returns
-    the job's id which is used to query status and retrieve results via get_job
+@decorators.api_view(['GET'])
+@decorators.permission_classes((AllowAny, ))
+def get_job(request, job_uuid, format=None):
+    # TODO consider if we should have some sort of session id check to ensure
+    # you can only view your own jobs.
+    try:
+        job = Job.objects.get(uuid=job_uuid)
+    except Job.DoesNotExist:
+        raise Http404("Not found.")
 
-    :param task_list: A list of Celery tasks to execute. Is made into a chain
-    :param job_input: Input to the first task, used in recording started jobs
-    :param user: The user requesting the job. Optional.
-    :return: A Response contianing the job id, marked as 'started'
-    """
-    created = now()
-    job = Job.objects.create(created_at=created, result='', error='',
-                             traceback='', user=user, status='started',
-                             model_input=job_input)
+    # Get the user so that logged in users can only see jobs that they started
+    # or anonymous ones
+    user = request.user if request.user.is_authenticated() else None
 
-    success = save_job_result.s(job.id, job_input)
-    error = save_job_error.s(job.id)
+    if job.user and job.user != user:
+        raise Http404("Not found.")
 
-    task_list.append(success)
-    task_chain = chain(task_list).apply_async(link_error=error)
-
-    job.uuid = task_chain.id
-    job.save()
-
+    # TODO Should we return the error? Might leak info about the internal
+    # workings that we don't want exposed.
     return Response(
         {
-            'job': task_chain.id,
-            'status': 'started',
+            'job_uuid': job.uuid,
+            'status': job.status,
+            'result': job.result,
+            'error': job.error,
+            'started': job.created_at,
+            'finished': job.delivered_at,
         }
     )
