+++ src/mmw/mmw/urls.py
@@ -12,6 +12,7 @@ import rest_framework.urls
 import apps.bigcz.urls
 import apps.geocode.urls
 import apps.modeling.urls
+import apps.geoprocessing_api.urls
 import apps.home.urls
 import apps.home.views
 import apps.water_balance.urls
@@ -31,6 +32,7 @@ urlpatterns = patterns(
     url(r'^api/bigcz/', include(apps.bigcz.urls)),
     url(r'^api/geocode/', include(apps.geocode.urls)),
     url(r'^api/modeling/', include(apps.modeling.urls)),
+    url(r'^api/geoprocessing/', include(apps.geoprocessing_api.urls)),
     url(r'^micro/', include(apps.water_balance.urls)),
     url(r'^user/', include(apps.user.urls))
 )
